resolving issues
