+++ src/main/java/com/bullhornsdk/data/model/entity/core/paybill/generalledger/GeneralLedgerSegment1.java
@@ -1,4 +1,4 @@
-package com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts;
+package com.bullhornsdk.data.model.entity.core.paybill.generalledger;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
@@ -11,10 +11,11 @@ import com.fasterxml.jackson.annotation.JsonRootName;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonRootName(value = "data")
 @JsonPropertyOrder({"id", "dateAdded", "dateLastModified", "externalSegmentName", "externalSegmentNumber", "generalLedgerSegmentType", "isActive", "isDeleted"})
-public class GeneralLedgerSegment4 extends GeneralLedgerSegment {
-    public GeneralLedgerSegment4() {
+public class GeneralLedgerSegment1 extends GeneralLedgerSegment {
+
+    public GeneralLedgerSegment1() {
         super();
-        this.setGeneralLedgerSegmentType(new GeneralLedgerSegmentType(4));
+        this.setGeneralLedgerSegmentType(new GeneralLedgerSegmentType(1));
     }
 
 }
