+++ src/main/java/com/bullhornsdk/data/model/entity/core/paybill/generalledger/GeneralLedgerServiceCode.java
@@ -1,4 +1,4 @@
-package com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts;
+package com.bullhornsdk.data.model.entity.core.paybill.generalledger;
 
 import com.bullhornsdk.data.model.entity.core.type.*;
 import com.bullhornsdk.data.util.ReadOnly;
