+++ src/main/java/com/bullhornsdk/data/model/entity/core/paybill/invoice/InvoiceTerm.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.entity.core.paybill.invoice;
 
-import com.bullhornsdk.data.model.entity.core.paybill.GeneralLedgerAccount;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerAccount;
 import com.bullhornsdk.data.model.entity.core.paybill.unit.CurrencyUnit;
 import com.bullhornsdk.data.model.entity.core.standard.ClientCorporation;
 import com.bullhornsdk.data.model.entity.core.type.*;
