--- src/main/java/com/bullhornsdk/data/model/entity/core/paybill/tax/TaxType.java
@@ -1,75 +1,0 @@
-package com.bullhornsdk.data.model.entity.core.paybill.tax;
-
-import com.bullhornsdk.data.model.entity.core.type.AbstractEntity;
-import com.bullhornsdk.data.model.entity.core.type.QueryEntity;
-import com.bullhornsdk.data.util.ReadOnly;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.JsonRootName;
-
-import java.util.Objects;
-
-/**
- * Created by mkesmetzis 23-Apr-20
- */
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonRootName(value = "data")
-@JsonPropertyOrder({"id", "label"})
-public class TaxType extends AbstractEntity implements QueryEntity {
-
-    private Integer id;
-    private String label;
-
-    public TaxType() {
-    }
-
-    public TaxType(Integer id) {
-        this.id = id;
-    }
-
-    @Override
-    @JsonProperty("id")
-    public Integer getId() {
-        return id;
-    }
-
-    @ReadOnly
-    @Override
-    @JsonProperty("id")
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    @JsonProperty("label")
-    public String getLabel() {
-        return label;
-    }
-
-    @JsonProperty("label")
-    public void setLabel(String label) {
-        this.label = label;
-    }
-
-    @Override
-    public String toString() {
-        return "TaxType{" +
-            "id=" + id +
-            ", label='" + label + '\'' +
-            '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        TaxType taxType = (TaxType) o;
-        return Objects.equals(id, taxType.id) &&
-            Objects.equals(label, taxType.label);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, label);
-    }
-}
