--- src/main/java/com/bullhornsdk/data/model/response/list/DiscountTypeListWrapper.java
@@ -1,11 +1,0 @@
-package com.bullhornsdk.data.model.response.list;
-
-import com.bullhornsdk.data.model.entity.core.paybill.discount.DiscountType;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({"data", "count", "start"})
-public class DiscountTypeListWrapper extends StandardListWrapper<DiscountType> {
-
-}
