+++ src/main/java/com/bullhornsdk/data/model/response/list/GeneralLedgerSegment1ListWrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.list;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerSegment1;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerSegment1;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
