+++ src/main/java/com/bullhornsdk/data/model/response/list/GeneralLedgerServiceCodeListWrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.list;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerServiceCode;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerServiceCode;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
