--- src/main/java/com/bullhornsdk/data/model/response/single/DiscountTypeWrapper.java
@@ -1,7 +1,0 @@
-package com.bullhornsdk.data.model.response.single;
-
-import com.bullhornsdk.data.model.entity.core.paybill.discount.DiscountType;
-
-public class DiscountTypeWrapper extends StandardWrapper<DiscountType> {
-
-}
