+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerAccountWrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.GeneralLedgerAccount;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerAccount;
 
 public class GeneralLedgerAccountWrapper extends StandardWrapper<GeneralLedgerAccount> {
