+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerSegment1Wrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerSegment1;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerSegment1;
 
 public class GeneralLedgerSegment1Wrapper extends StandardWrapper<GeneralLedgerSegment1> {
