+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerSegment2Wrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerSegment2;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerSegment2;
 
 public class GeneralLedgerSegment2Wrapper extends StandardWrapper<GeneralLedgerSegment2> {
