+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerSegment3Wrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerSegment3;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerSegment3;
 
 public class GeneralLedgerSegment3Wrapper extends StandardWrapper<GeneralLedgerSegment3> {
