+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerSegment4Wrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerSegment4;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerSegment4;
 
 public class GeneralLedgerSegment4Wrapper extends StandardWrapper<GeneralLedgerSegment4> {
