+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerSegment5Wrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerSegment5;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerSegment5;
 
 public class GeneralLedgerSegment5Wrapper extends StandardWrapper<GeneralLedgerSegment5> {
