+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerSegmentTypeWrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerSegmentType;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerSegmentType;
 
 public class GeneralLedgerSegmentTypeWrapper extends StandardWrapper<GeneralLedgerSegmentType> {
