+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerSegmentWrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerSegment;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerSegment;
 
 public class GeneralLedgerSegmentWrapper extends StandardWrapper<GeneralLedgerSegment> {
