+++ src/main/java/com/bullhornsdk/data/model/response/single/GeneralLedgerServiceCodeWrapper.java
@@ -1,6 +1,6 @@
 package com.bullhornsdk.data.model.response.single;
 
-import com.bullhornsdk.data.model.entity.core.paybill.chartofaccounts.GeneralLedgerServiceCode;
+import com.bullhornsdk.data.model.entity.core.paybill.generalledger.GeneralLedgerServiceCode;
 
 public class GeneralLedgerServiceCodeWrapper extends StandardWrapper<GeneralLedgerServiceCode> {
